requirements.txt
flask==3.1.1
flask-sqlalchemy==3.1.1
psycopg2==2.9.11
python-dotenv==1.0.0
flask-bcrypt==1.0.1
flask-cors==4.0.0
google-generativeai>=0.8.0
flask-swagger-ui==4.11.1